move stores/indexes creation to schema
