+++ lib/idb-store.js
@@ -10,13 +10,12 @@ module.exports = Store;
 /**
  * Initialize new `Store`.
  *
- * @param {Treo} db
  * @param {String} name
  * @param {Object} opts
  */
 
-function Store(db, name, opts) {
-  this.db = db;
+function Store(name, opts) {
+  this.db = null;
   this.name = name;
   this.indexes = {};
   this.key = opts.key || opts.keyPath || undefined;
