+++ lib/index.js
@@ -25,18 +25,13 @@ function Treo(name, schema) {
   this.name = name;
   this.status = 'close';
   this.origin = null;
-  this.stores = {};
+  this.stores = schema._stores;
   this.version = schema.getVersion();
-  this.versions = schema.getVersions();
+  this.onupgradeneeded = schema.callback();
 
-  // setup stores and indexes
-  Object.keys(schema._stores).forEach(function(storeName) {
-    var s = schema._stores[storeName];
-    var store = this.stores[storeName] = new Store(this, s.name, s.opts);
-    Object.keys(s.indexes).forEach(function(indexName) {
-      var i = s.indexes[indexName];
-      store.indexes[indexName] = new Index(store, i.name, i.field, i.opts);
-    });
+  // assign db property to each store
+  Object.keys(this.stores).forEach(function(storeName) {
+    this.stores[storeName].db = this;
   }, this);
 }
 
@@ -123,30 +118,7 @@ Treo.prototype.getInstance = function(cb) {
 
   var that = this;
   var req = indexedDB().open(this.name, this.version);
-
-  req.onupgradeneeded = function onupgradeneeded(e) {
-    var db = e.target.result;
-    var tr = e.target.transaction;
-
-    that.versions.forEach(function(versionSchema) {
-      if (e.oldVersion >= versionSchema.version) return;
-
-      versionSchema.stores.forEach(function(s) {
-        db.createObjectStore(s.name, {
-          keyPath: s.opts.key || s.opts.keyPath,
-          autoIncrement: s.opts.increment || s.opts.autoIncrement,
-        });
-      });
-
-      versionSchema.indexes.forEach(function(i) {
-        var store = tr.objectStore(i.store.name);
-        store.createIndex(i.name, i.field, {
-          unique: i.opts.unique,
-          multiEntry: i.opts.multi || i.opts.multiEntry,
-        });
-      });
-    });
-  };
+  req.onupgradeneeded = this.onupgradeneeded;
 
   req.onerror = req.onblocked = function onerror(e) {
     that.status = 'error';
