+++ lib/schema.js
@@ -1,4 +1,6 @@
 var type = require('component-type');
+var Store = require('./idb-store');
+var Index = require('./idb-index');
 
 /**
  * Expose `Schema`.
@@ -12,8 +14,8 @@ module.exports = Schema;
 
 function Schema() {
   if (!(this instanceof Schema)) return new Schema();
-  this._current = {};
   this._stores = {};
+  this._current = {};
   this._versions = {};
 }
 
@@ -44,9 +46,9 @@ Schema.prototype.version = function(version) {
 Schema.prototype.addStore = function(name, opts) {
   if (type(name) != 'string') throw new TypeError('`name` is required');
   if (this._stores[name]) throw new TypeError('store is already defined');
-  var store = { name: name, indexes: {}, opts: opts || {} };
-  this._versions[this.getVersion()].stores.push(store);
+  var store = new Store(name, opts || {});
   this._stores[name] = store;
+  this._versions[this.getVersion()].stores.push(store);
   this._current.store = store;
   return this;
 };
@@ -64,8 +66,8 @@ Schema.prototype.addIndex = function(name, field, opts) {
   if (type(name) != 'string') throw new TypeError('`name` is required');
   if (type(field) != 'string') throw new TypeError('`field` is required');
   var store = this._current.store;
-  var index = { store: store, name: name, field: field, opts: opts || {} };
   if (store.indexes[name]) throw new TypeError('index is already defined');
+  var index = new Index(store, name, field, opts || {});
   store.indexes[name] = index;
   this._versions[this.getVersion()].indexes.push(index);
   return this;
@@ -96,13 +98,37 @@ Schema.prototype.getVersion = function() {
 };
 
 /**
- * Get sorted array of versions.
+ * Generate onupgradeneeded callback.
  *
- * @return {Array}
+ * @return {Function}
  */
 
-Schema.prototype.getVersions = function() {
-  return Object.keys(this._versions)
+Schema.prototype.callback = function() {
+  var versions = Object.keys(this._versions)
     .map(function(v) { return this._versions[v] }, this)
     .sort(function(a, b) { return a.version - b.version });
+
+  return function onupgradeneeded(e) {
+    var db = e.target.result;
+    var tr = e.target.transaction;
+
+    versions.forEach(function(versionSchema) {
+      if (e.oldVersion >= versionSchema.version) return;
+
+      versionSchema.stores.forEach(function(s) {
+        db.createObjectStore(s.name, {
+          keyPath: s.key,
+          autoIncrement: s.increment
+        });
+      });
+
+      versionSchema.indexes.forEach(function(i) {
+        var store = tr.objectStore(i.store.name);
+        store.createIndex(i.name, i.field, {
+          unique: i.unique,
+          multiEntry: i.multi
+        });
+      });
+    });
+  };
 };
